+++ templates/login.html
@@ -6,6 +6,15 @@
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
     <link rel="stylesheet" href="{{url_for('static', path='/css/login.css')}}" type="text/css">
     <meta name="viewport" content="width=device-width, initial-scale=1">
+    <script>
+        if (window.history && window.history.replaceState) {
+            window.history.replaceState(null, "", window.location.href);
+            window.history.pushState(null, "", window.location.href);
+            window.onpopstate = function () {
+                window.location.href = "/";
+            };
+        }
+    </script>
 </head>
 
 <body>
